refactor: rename tasks-calendar provider to context
Replaced `provider` with `context` in tasks-calendar components for consistency with established naming conventions. Updated all imports, exports, and references accordingly.
