+++ apps/nextjs/src/components/features/tasks/tasks-calendar/tasks-calendar.tsx
@@ -1,6 +1,6 @@
 import { forwardRef, memo } from 'react';
 import { Flex, type FlexProps } from '@/components/ui/flex';
-import { Context } from './provider';
+import { Context } from './context';
 
 type Props = FlexProps;
