+++ apps/nextjs/src/components/features/tasks/tasks-header/calendar-month-picker/calendar-month-picker.tsx
@@ -1,5 +1,5 @@
 import { memo, useMemo } from 'react';
-import { useTasksCalendarContext } from '@/components/features/tasks/tasks-calendar/provider';
+import { useTasksCalendarContext } from '@/components/features/tasks/tasks-calendar/context';
 import { Flex } from '@/components/ui/flex';
 import { Icon } from '@/components/ui/icon';
 import { IconButton } from '@/components/ui/icon-button';
